Add Ex13-5
Consider the application of nearest-neighbors to the easy and hard problems in the left panel of Figure 13.5. 1. Replicate the results in the left panel of Figure 13.5. 2. Estimate the misclassification errors using fivefold cross-validation, and compare the error rate curves to those in 1. 3. Consider an “AIC-like” penalization of the training set misclassification error. Specifically, add 2t/N to the training set misclassification error, where t is the approximate number of parameters N/r, r being the number of nearest-neighbors. Compare plots of the resulting penalized misclassification error to those in 1 and 2. Which method gives a better estimate of the optimal number of nearest-neighbors: cross-validation or AIC?